Support refs to unsized SqlTranslatable types (#1890)
This properly separates the question of "does this have an SQL
representation?" from "is this a viable function argument?" While it has
no immediate benefit aside from as a minor cleanup, it will prove useful
in supporting types and functions where an unsized type is used as a
generic. Such code will return slice references and the like.