`cargo pgrx test --runas` envar passing (#1674)
This teaches the test `framework.rs` to pass through to `sudo` all the
various cargo/rust environment variables.

Turns out this might be necessary for `#[pg_test]` tests that use
`std::env::var(...)`. When using `--runas` we spawn the postmaster with
`sudo`, and its default policy is to reset the environment (`env_reset`
flag in the sudoers file).