Fixed docker file for alpine (#2024)
Previously I made a pull request that should fix the CI, but I forgot to
include changes of the docker file. So, now it includes the changes and
in addition to that the file name doesn't include version anymore to
make it a bit easier to maintain. Only the docker file needs to be
updated and a newer version set in it.

The issue with the Fedora's part is quite simple. PostgreSQL 12 is dead
and no more supported, that's why it doesn't exist in the repository and
cannot be installed on Fedora with the current docker file.

See [my CI
run](https://github.com/YohDeadfall/pgrx/actions/runs/14229043458) that
verifies that the fix works.